frontend for the Home page , Login and Sign up pages has been added
For now I just have attached the frontends of homepage , login and sign up pages . I am working on the backend for these pages with Django. As soon as I am done with this work I will join all the pages and complete the website. Thereafter we can host it on Azure and add the ML feature to the website.